requirements.txt
gspread
google-auth-oauthlib
google-auth-httplib2
pandas
playwright
python-dotenv
ollama